+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/ConsoleUnit.java
@@ -20,11 +20,6 @@ public interface ConsoleUnit {
         return getTexture(entity, false);
     }
 
-
-
     ResourceLocation getDefaultTexture();
-
-    ResourceLocation getConsoleTheme();
-
 
 }