+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/CoralConsoleModel.java
@@ -72,11 +72,6 @@ public class CoralConsoleModel extends HierarchicalModel implements ConsoleUnit 
     }
 
     @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.CORAL.getId();
-    }
-
-    @Override
     public ResourceLocation getDefaultTexture() {
         return CORAL_TEXTURE;
     }