+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/FactoryConsoleModel.java
@@ -112,9 +112,4 @@ public class FactoryConsoleModel extends HierarchicalModel implements ConsoleUni
     public ResourceLocation getDefaultTexture() {
         return FACTORY_TEXTURE;
     }
-
-    @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.FACTORY.getId();
-    }
 }