+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/InitiativeConsoleModel.java
@@ -82,8 +82,4 @@ public class InitiativeConsoleModel extends HierarchicalModel implements Console
         return INITIATIVE_TEXTURE;
     }
 
-    @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.INITIATIVE.getId();
-    }
 }