+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/MystConsoleModel.java
@@ -91,8 +91,4 @@ public class MystConsoleModel extends HierarchicalModel implements ConsoleUnit {
         return MYST_TEXTURE;
     }
 
-    @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.MYST.getId();
-    }
 }