+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/ToyotaConsoleModel.java
@@ -80,8 +80,4 @@ public class ToyotaConsoleModel extends HierarchicalModel implements ConsoleUnit
         return TOYOTA_TEXTURE;
     }
 
-    @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.TOYOTA.getId();
-    }
 }