+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/console/VictorianConsoleModel.java
@@ -111,8 +111,4 @@ public class VictorianConsoleModel extends HierarchicalModel implements ConsoleU
         return VICTORIAN_TEXTURE;
     }
 
-    @Override
-    public ResourceLocation getConsoleTheme() {
-        return ConsoleTheme.VICTORIAN.getId();
-    }
 }