+++ common/src/main/java/whocraft/tardis_refined/client/renderer/blockentity/device/ConsoleConfigurationRenderer.java
@@ -50,7 +50,7 @@ public class ConsoleConfigurationRenderer implements BlockEntityRenderer<Console
 
             ResourceLocation theme = blockEntity.theme();
 
-            ConsoleUnit consoleModel = ConsoleModelCollection.getInstance().getConsoleModel(theme);
+            ConsoleUnit consoleModel = ConsoleModelCollection.getInstance().getConsoleEntry(theme).getConsoleModel(null);
             if (consoleModel != null) {
                 consoleModel.renderConsole(null, blockEntity.getLevel(), poseStack, bufferSource.getBuffer(RenderType.entityTranslucent(consoleModel.getDefaultTexture())), packedLight, OverlayTexture.NO_OVERLAY, 0.635f, 0.392f, 0.878f, 0.5f);
             }