+++ forge/src/main/java/whocraft/tardis_refined/common/data/ProviderBlockTags.java
@@ -1,5 +1,6 @@
 package whocraft.tardis_refined.common.data;
 
+import com.simibubi.create.content.decoration.palettes.GlassPaneBlock;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
@@ -94,7 +95,7 @@ public class ProviderBlockTags extends BlockTagsProvider {
         Set<Block> glassBlocks = new HashSet<>();
         ManipulatorRecipes.MANIPULATOR_CRAFTING_RECIPES.forEach((resourceLocation, manipulatorCraftingRecipe) -> {
             for (ManipulatorCraftingIngredient ingredient : manipulatorCraftingRecipe.ingredients()) {
-                if(ingredient.inputBlockState().getBlock() instanceof GlassBlock) {
+                if(ingredient.inputBlockState().getBlock() instanceof GlassBlock || ingredient.inputBlockState().getBlock() instanceof GlassPaneBlock || ingredient.inputBlockState().getBlock() instanceof GlassPaneBlock) {
                     glassBlocks.add(ingredient.inputBlockState().getBlock());
                 }
             }